requirements.txt
streamlit
faiss-cpu
sentence-transformers
transformers
torch
pillow
pytesseract
python-docx
PyPDF2
faster-whisper
ffmpeg-python
groq
numpy
